% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calculate_delta_method}
\alias{calculate_delta_method}
\title{Calculate delta method estimates for Cmax}
\usage{
calculate_delta_method(mod1, config)
}
\arguments{
\item{mod1}{Fitted SAEMIX object}

\item{config}{Model configuration}
}
\value{
Data frame with delta method results
}
\description{
Calculate delta method estimates for Cmax
}
\keyword{internal}
