% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate quantiles for different TOST approaches}
\usage{
calculate_quantiles(n, nb_t, nb_param)
}
\arguments{
\item{n}{Number of subjects}

\item{nb_t}{Number of timepoints}

\item{nb_param}{Number of parameters}
}
\value{
List with quantiles
}
\description{
Calculate quantiles for different TOST approaches
}
\keyword{internal}
