% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{fit_saemix_model}
\alias{fit_saemix_model}
\title{Fit SAEMIX model}
\usage{
fit_saemix_model(data, config, nb_chains, nb_iter)
}
\arguments{
\item{data}{Preprocessed data frame}

\item{config}{Model configuration}

\item{nb_chains}{Number of MCMC chains}

\item{nb_iter}{Vector of burn-in and sampling iterations}
}
\value{
Fitted SAEMIX object
}
\description{
Fit SAEMIX model
}
\keyword{internal}
