% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_model_config}
\alias{get_model_config}
\title{Get model configuration}
\usage{
get_model_config(model_type)
}
\arguments{
\item{model_type}{Character string specifying the PK model}
}
\value{
List with model configuration
}
\description{
Get model configuration
}
\keyword{internal}
