% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{MNLEM-package}
\alias{MNLEM}
\alias{MNLEM-package}
\title{MNLEM: Analyse de Bioéquivalence avec Modèles Non-Linéaires à Effets Mixtes}
\description{
Ce package effectue des analyses de bioéquivalence sur des paramètres pharmacocinétiques à l'aide de modèles non-linéaires à effets mixtes (MNLEM). Il implémente la méthode TOST (Two One-Sided Tests) pour tester l'équivalence des paramètres AUC et Cmax.
}
\author{
\strong{Maintainer}: Eden BETE \email{eden.bete.etu@univ-lille.fr}

Authors:
\itemize{
  \item Celia BENGUENNA \email{celia.benguenna.etu@univ-lille.fr}
}

}
\keyword{internal}
