% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{model_1cpt}
\alias{model_1cpt}
\title{One-compartment model with first-order absorption}
\usage{
model_1cpt(psi, id, xidep)
}
\description{
One-compartment model with first-order absorption
}
\keyword{internal}
