% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{model_1cpt_tlag}
\alias{model_1cpt_tlag}
\title{One-compartment model with first-order absorption and lag time}
\usage{
model_1cpt_tlag(psi, id, xidep)
}
\description{
One-compartment model with first-order absorption and lag time
}
\keyword{internal}
