% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{model_2cpt}
\alias{model_2cpt}
\title{Two-compartment model with first-order absorption}
\usage{
model_2cpt(psi, id, xidep)
}
\description{
Two-compartment model with first-order absorption
}
\keyword{internal}
