% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{model_2cpt_tlag}
\alias{model_2cpt_tlag}
\title{Two-compartment model with first-order absorption and lag time}
\usage{
model_2cpt_tlag(psi, id, xidep)
}
\description{
Two-compartment model with first-order absorption and lag time
}
\keyword{internal}
