% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{perform_tost_auc}
\alias{perform_tost_auc}
\title{Perform TOST tests for AUC}
\usage{
perform_tost_auc(mod1, config, quantiles, n, nb_param)
}
\arguments{
\item{mod1}{Fitted SAEMIX object}

\item{config}{Model configuration}

\item{quantiles}{List of quantiles}

\item{n}{Number of subjects}

\item{nb_param}{Number of parameters}
}
\value{
List with AUC TOST results
}
\description{
Perform TOST tests for AUC
}
\keyword{internal}
