% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{perform_tost_cmax}
\alias{perform_tost_cmax}
\title{Perform TOST tests for Cmax}
\usage{
perform_tost_cmax(SE_deltam, quantiles, n, nb_param)
}
\arguments{
\item{SE_deltam}{Delta method results}

\item{quantiles}{List of quantiles}

\item{n}{Number of subjects}

\item{nb_param}{Number of parameters}
}
\value{
List with Cmax TOST results
}
\description{
Perform TOST tests for Cmax
}
\keyword{internal}
