% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print_analysis_summary}
\alias{print_analysis_summary}
\title{Print analysis summary}
\usage{
print_analysis_summary(results, model_type, data_path, n, nb_t)
}
\arguments{
\item{results}{TOST results}

\item{model_type}{Model type}

\item{data_path}{Data path}

\item{n}{Number of subjects}

\item{nb_t}{Number of timepoints}
}
\description{
Print analysis summary
}
\keyword{internal}
