% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{stable_diff}
\alias{stable_diff}
\title{Stable difference function to prevent numerical issues}
\usage{
stable_diff(a, b, min_diff = 1e-06)
}
\description{
Stable difference function to prevent numerical issues
}
\keyword{internal}
