% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tost_bioequivalence_analysis}
\alias{tost_bioequivalence_analysis}
\title{Fonction principale pour l'analyse de bioéquivalence (TOST) sur AUC et Cmax}
\usage{
tost_bioequivalence_analysis(
  filepath = NULL,
  delta = log(1.25),
  nb_chains = 10,
  nb_iter = c(300, 100)
)
}
\arguments{
\item{filepath}{Chemin vers le fichier de données contenant les observations (par défaut `"dataset_1.txt"`).}

\item{delta}{Seuil d’équivalence (logarithmique), généralement \code{log(1.25)}. Ce seuil est utilisé pour définir
les marges de bioéquivalence.}

\item{nb_chains}{Nombre de chaînes pour l’algorithme SAEM (par défaut \code{10}).}

\item{nb_iter}{Nombre d'itérations pour l’algorithme SAEM, spécifié sous forme de vecteur (par défaut \code{c(300, 100)}).}
}
\value{
Aucun objet retourné. Les résultats des tests TOST sont affichés dans la console.
}
\description{
Cette fonction effectue les tests de bioéquivalence (TOST) sur les paramètres pharmacocinétiques log(AUC) et log(Cmax)
pour une étude de bioéquivalence à deux bras parallèles. Le modèle utilisé est un modèle pharmacocinétique à un
compartiment ajusté avec l'algorithme SAEM via le package `saemix`.
}
\examples{
tost_bioequivalence_analysis()
}
