% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{validate_and_load_data}
\alias{validate_and_load_data}
\title{Validate and preprocess input data}
\usage{
validate_and_load_data(data_path)
}
\arguments{
\item{data_path}{Path to the dataset file}
}
\value{
Preprocessed data frame
}
\description{
Validate and preprocess input data
}
\keyword{internal}
